services(openai): rename OpenAILLMServiceRealtimeBeta to OpenAIRealtimeBetaLLMService
