+++ src/pipecat/services/openai_realtime_beta/__init__.py
@@ -1,2 +1,2 @@
 from .events import InputAudioTranscription, SessionProperties, TurnDetection
-from .llm_and_context import OpenAILLMServiceRealtimeBeta
+from .openai import OpenAIRealtimeBetaLLMService
